requirements.txt
Flask==2.2.5
Flask_Cors==4.0.1
mysql_connector_repackaged==0.3.1
pandas==2.2.2
scikit_learn==1.2.2
